fix: [samba] when lost netconnection, computer page stocked
when try to access samba mounts via computer/sidebar item while network
is not accessable, computer page stocked for a longlong time.
make a 2sec (defalut) timeout when access mounts in this situation, and
when timeouted, raise a 'cannot access' error dialog.

Log: fix issue.

Bug: https://pms.uniontech.com/bug-view-128039.html

+++ src/dfm-base/utils/dialogmanager.cpp
@@ -103,7 +103,7 @@ void DialogManager::showErrorDialogWhenMountDeviceFailed(DFMMOUNT::DeviceError e
         break;
     default:
         showErrorDialog(tr("Mount error"), tr("Error occured while mounting device"));
-        qWarning() << "mount device failed: " << static_cast<int>(err);
+        qWarning() << "mount device failed: " << err;
         break;
     }
 }
@@ -364,7 +364,7 @@ void DialogManager::showRestoreFailedDialog(const QList<QUrl> &urlList)
         d.setMessage(tr("Failed to restore %1 files, the target folder is read-only").arg(QString::number(urlList.count())));
     }
     d.setIcon(QIcon::fromTheme("dialog-warning"));
-    d.addButton(tr("OK","button"), true, DDialog::ButtonRecommend);
+    d.addButton(tr("OK", "button"), true, DDialog::ButtonRecommend);
     d.exec();
 }
